+++ src/ptex/PtexTriangleFilter.h
@@ -67,8 +67,8 @@ class PtexTriangleFilter : public PtexFilter, public Ptex
     
     PtexTexture* _tx;		// texture being evaluated
     Options _options;		// options
-    double* _result;		// temp result
-    double _weight;		// accumulated weight of data in _result
+    float* _result;		// temp result
+    float _weight;		// accumulated weight of data in _result
     int _firstChanOffset;	// byte offset of first channel to eval
     int _nchan;			// number of channels to eval
     int _ntxchan;		// number of channels in texture
